% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccxt-package.R
\docType{package}
\name{ccxt-package}
\alias{ccxt}
\alias{ccxt-package}
\title{ccxt: ccxt Analysis}
\description{
| Analysis of crypto markets via `ccxt`.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JohnGavin/ccxt}
  \item Report bugs at \url{https://github.com/JohnGavin/ccxt/issues}
}

}
\author{
\strong{Maintainer}: John Gavin \email{john@1x2.ltd}

Other contributors:
\itemize{
  \item ESBH [copyright holder]
}

}
\keyword{internal}
